2510_04618/pyproject.toml
[project]
name = "2510-04618"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.16.0",
]
